+++ NEWS.md
@@ -2,12 +2,14 @@
 
 * Changes in recommendation method AR: Default for maxlen is now 3 to 
     find more specific rules. Parameters measure and decreasing for 
-    sorting the rule baseare now called sort_measure and sort_decreasing.
+    sorting the rule base are now called sort_measure and sort_decreasing.
     New parameter apriori_control can be used to pass a control list to
     apriori in arules.
 * The registry now has a reference field.
-* Added recommender method ALS and ALS_implicit nased on latent factors 
+* Added recommender method ALS and ALS_implicit based on latent factors 
     and alternating least squares (contributed by Bregt Verreet).
+* Fixed bug in method IBCF with n being ignored in 
+    predict (reported by Giorgio Alfredo Spedicato).
 
 # Changes in version 0.2-0 (05/31/2016)